pyproject.toml
[project]
name = "gridfinity-designer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "numpy-stl>=3.2.0",
    "pydantic>=2.13.3",
    "uvicorn[standard]>=0.46.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "trimesh>=4.12.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
